Add support for checkpoint and restore via CRIU
This commit adds support for checkpointing and restoring containers via CRIU.
It adds two api calls, ->checkpoint and ->restore, which are wrappers around
the CRIU CLI. CRIU has an RPC API, but reasons for preferring exec() are
discussed in [1].

To checkpoint, users specify a directory to dump the container metadata (CRIU
dump files, plus some additional information about veth pairs and which bridges
they are attached to) into this directory. On restore, this information is read
out of the directory, a CRIU command line is constructed, and CRIU is exec()d.
Criu uses the lxc-restore-net callback (which in turn inspects the image
directory with the NIC data) to properly restore the network.

[1]: http://lists.openvz.org/pipermail/criu/2014-July/015117.html

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>