support use of 'all' containers when cgmanager supports it (v2)
This also fixes an apparent existing bug where the cgmanager
connection was not disconnected on failure.

Introduce a new list of controllers just containing "all".

Make the lists of controllers null-terminated.

If the cgmanager api version is high enough, use the 'all' controller
rather than walking all controllers, which should greatly reduce the
amount of dbus overhead.  This will be especially important for
those going through a cgproxy.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>