logs: introduce a thread-local 'current' lxc_config
If the lxc container configuration specifies a loglevel/logfile, only
set the lxc_config's logfd and loglevel according to those.

Each API call is wrapped to set/unset the current_config.

Update logfile appender to use the logfile specified in lxc_conf if (a)
current_config is set and (b) the lxc-* command did not override it.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>