cgmangaer: don't ask to switch to a cgroup if we're already in it
If we are in /user.slice/user-1000.slice/session-c2.scope, and we
start an unprivileged container t1, it will be in cgroup
3:memory:/user.slice/user-1000.slice/session-c2.scope/t1.  If
we then do a 'lxc-cgroup -n t1 freezer.tasks', cgm_get will
first switch to 3:memory:/user.slice/user-1000.slice/session-c2.scope
then look up 't1's values.  The reasons for this are

1. cgmanager get_value is relative to your own cgroup, so we need
to be sure to be in t1's cgroup or an ancestor
2. we don't want to be in the container's cgroup bc it might freeze us.

But in Ubuntu 15.04 it was decided that
3:memory:/user.slice/user-1000.slice/session-c2.scope/tasks should
not be writeable by the user, making this fail.  Therefore, if we
are already in 3:memory:/user.slice/user-1000.slice/session-c2.scope
then don't try to switch to it.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>