cgroups: work around inability to start whitelist
We cannot switch from devices cgroup blacklist to whitelist if
there are sub-cgroups.  With unified hierarchy tasks must be in
leaf nodes so we must create a sub-cgroup to move the container
init into.  This causes a chicken and egg problem - we can't
move the container task into the cgroup later than we are, because
unprivileged users won't have the privilege.  We cannot blindly
set the container whitelist earlier because privileged containers
won't be able to mount devices.

So switch to a whitelist (if needed) early on, but also set
whitelist entries covering everything.  These will be cleared
out later by the rule which would otherwise be switching to whitelist,
'devices.deny = a'.

Signed-off-by: Serge Hallyn <serge.hallyn@ubuntu.com>